Rename feed_components into components
